cleanup server code
